Added the templated class OpenCL_Array<T>::OpenCL_Array().
This allows managing an array on host and device memory transparently.
Initializing the object will allocate memory on the device. Functions to
transfer from host to device and device to host are present.

Additionally, checksums can be used to verify to transfer at any time
using Validate_Data(). This perform an SHA512 checksum on both the host
and device memory and compare (assert) the two values. Note that
checksumming requires padding. If enabled, the host array WILL be
reallocated to make room for the padding.
To use, the checksumming, make sure you define "OpenCLSHA512Checksum" at
compile time. If using the cmake, uncomment CMakeLists.txt's line 11.